refactor(chatbot-config): split client logic into separate component for SSR compatibility
