+++ ui/components/lighthouse/index.ts
@@ -1,0 +1,2 @@
+export * from "./chat";
+export * from "./chatbot-config";
